requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.8.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-language==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.20.4
huggingface-hub==0.10.1
idna==3.4
numpy==1.23.4
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tokenizers==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
